all reagester controler

+++ src/utils/cloudinary.js
@@ -17,12 +17,17 @@ const uploadOnCloudinary = async (localFilePath) => {
             resource_type:"auto"
         })
         // file has been uploaded successfull
-        console.log("file is uploaded on cloudinary" , response.url);
+        // console.log("file is uploaded on cloudinary" , response.url);
+        fs.unlinkSync(localFilePath)
         return response;
 
     }catch(error){
-        fs.unlinkSync(localFilePath)  // remove the locally saved tempaorary file as the upload operation got failed
-        return null;
+        console.error("Cloudinary Upload Error:", error);
+
+    if (localFilePath && fs.existsSync(localFilePath)) {
+        fs.unlinkSync(localFilePath);
+    }
+
     }
 }
